Move stats to metrics
