+++ api/src/main/java/openconsensus/metrics/GaugeDouble.java
@@ -42,7 +42,7 @@ import openconsensus.metrics.GaugeDouble.TimeSeries;
  *
  *   void doWork() {
  *      // Your code here.
- *      defaultPoint.add(10);
+ *      defaultTimeSeries.add(10);
  *   }
  *
  * }
@@ -67,7 +67,7 @@ import openconsensus.metrics.GaugeDouble.TimeSeries;
  *
  *   void doSomeWork() {
  *      // Your code here.
- *      inboundPoint.set(15);
+ *      inboundTimeSeries.set(15);
  *   }
  *
  * }
