+++ api/src/main/java/openconsensus/metrics/Measure.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package openconsensus.stats;
+package openconsensus.metrics;
 
 import javax.annotation.concurrent.ThreadSafe;
