+++ api/src/main/java/openconsensus/metrics/Measurement.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package openconsensus.stats;
+package openconsensus.metrics;
 
 import javax.annotation.concurrent.Immutable;
