--- api/src/main/java/openconsensus/stats/Stats.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2019, OpenConsensus Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package openconsensus.stats;
-
-/**
- * Class for accessing the global {@link StatsRecorder}.
- *
- * @since 0.1.0
- */
-public final class Stats {
-  private static final StatsRecorder STATS_RECORDER = NoopStats.newNoopStatsRecorder();
-
-  /**
-   * Returns the global {@link StatsRecorder}.
-   *
-   * @return stats recorder.
-   * @since 0.1.0
-   */
-  public static StatsRecorder getStatsRecorder() {
-    return STATS_RECORDER;
-  }
-
-  private Stats() {}
-}
